refactor/rename pockydb
