+++ src/lib/response-triggers/winners.ts
@@ -1,7 +1,7 @@
 import Trigger from '../../models/trigger';
 import constants from '../../constants';
 import TableHelper from '../parsers/tableHelper';
-import PockyDB from '../PockyDB';
+import PockyDB from '../database/pocky-db';
 import Config from '../config';
 import __logger from '../logger';
 import { MessageObject } from 'ciscospark/env';
